refactor: move store_thumbnail function to ayon_server.helpers.thumbnails
